Consistent rename of fragments / tests to test_cases
