+++ contrib/automation_tests/orbit_bottom_up.py
@@ -7,9 +7,9 @@ found in the LICENSE file.
 from absl import app
 
 from core.orbit_e2e import E2ETestSuite
-from fragments.connection_window import FilterAndSelectFirstProcess, ConnectToStadiaInstance
-from fragments.capture_window import Capture
-from fragments.bottom_up_tab import VerifyHelloGgpBottomUpContents
+from test_cases.connection_window import FilterAndSelectFirstProcess, ConnectToStadiaInstance
+from test_cases.capture_window import Capture
+from test_cases.bottom_up_tab import VerifyHelloGgpBottomUpContents
 
 """Inspect the bottom-up view in Orbit using pywinauto.
 
@@ -32,13 +32,13 @@ This automation script covers a basic workflow:
 
 
 def main(argv):
-    tests = [
+    test_cases = [
         ConnectToStadiaInstance(),
         FilterAndSelectFirstProcess(process_filter='hello_ggp'),
         Capture(),
         VerifyHelloGgpBottomUpContents()
     ]
-    suite = E2ETestSuite(test_name="Bottom-Up View", tests=tests)
+    suite = E2ETestSuite(test_name="Bottom-Up View", test_cases=test_cases)
     suite.execute()
