+++ contrib/automation_tests/orbit_connect_and_capture.py
@@ -7,17 +7,17 @@ found in the LICENSE file.
 from absl import app
 
 from core.orbit_e2e import E2ETestSuite
-from fragments.connection_window import FilterAndSelectFirstProcess, ConnectToStadiaInstance
-from fragments.capture_window import Capture
+from test_cases.connection_window import FilterAndSelectFirstProcess, ConnectToStadiaInstance
+from test_cases.capture_window import Capture
 
 
 def main(argv):
-    tests = [
+    test_cases = [
         ConnectToStadiaInstance(),
         FilterAndSelectFirstProcess(process_filter='hello_'),
         Capture()
     ]
-    suite = E2ETestSuite(test_name="Connect & Capture", tests=tests)
+    suite = E2ETestSuite(test_name="Connect & Capture", test_cases=test_cases)
     suite.execute()
