+++ contrib/automation_tests/orbit_move_tabs.py
@@ -7,18 +7,18 @@ found in the LICENSE file.
 from absl import app, flags
 
 from core.orbit_e2e import E2ETestSuite
-from fragments.connection_window import ConnectToStadiaInstance, FilterAndSelectFirstProcess
-from fragments.main_window import MoveTab
+from test_cases.connection_window import ConnectToStadiaInstance, FilterAndSelectFirstProcess
+from test_cases.main_window import MoveTab
 
 
 def main(argv):
     tab_title = "Symbols" if flags.FLAGS.enable_ui_beta else "Functions"
-    tests = [
+    test_cases = [
         ConnectToStadiaInstance(),
         FilterAndSelectFirstProcess(process_filter="hello_ggp"),
         MoveTab(tab_title=tab_title, tab_name="FunctionsTab")
     ]
-    suite = E2ETestSuite(test_name="Move tabs", tests=tests)
+    suite = E2ETestSuite(test_name="Move tabs", test_cases=test_cases)
     suite.execute()
