+++ contrib/automation_tests/orbit_top_down.py
@@ -7,9 +7,9 @@ found in the LICENSE file.
 from absl import app
 
 from core.orbit_e2e import E2ETestSuite
-from fragments.connection_window import FilterAndSelectFirstProcess, ConnectToStadiaInstance
-from fragments.capture_window import Capture
-from fragments.top_down_tab import VerifyHelloGgpTopDownContents
+from test_cases.connection_window import FilterAndSelectFirstProcess, ConnectToStadiaInstance
+from test_cases.capture_window import Capture
+from test_cases.top_down_tab import VerifyHelloGgpTopDownContents
 
 """Inspect the top-down view in Orbit using pywinauto.
 
@@ -33,13 +33,13 @@ This automation script covers a basic workflow:
 
 
 def main(argv):
-    tests = [
+    test_cases = [
         ConnectToStadiaInstance(),
         FilterAndSelectFirstProcess(process_filter='hello_ggp'),
         Capture(),
         VerifyHelloGgpTopDownContents()
     ]
-    suite = E2ETestSuite(test_name="Top-Down View", tests=tests)
+    suite = E2ETestSuite(test_name="Top-Down View", test_cases=test_cases)
     suite.execute()
