+++ contrib/automation_tests/test_cases/bottom_up_tab.py
