+++ contrib/automation_tests/test_cases/capture_window.py
