+++ contrib/automation_tests/test_cases/connection_window.py
