+++ contrib/automation_tests/test_cases/main_window.py
